requirements-test.txt
# test runs requirements (versions we'll be testing against) - automatically updated by requires.io
coverage==4.1 # pytest-cov
mirakuru==0.8.2
mysqlclient==1.3.7
path.py==8.2.1
port-for==0.3.1
psycopg2==2.6.2
pyaml==15.8.2
pymlconf==0.4.0
pymongo==3.3.0
pytest==2.8.7
rabbitpy==0.27.1
redis==2.10.5
pyflakes==1.2.3
pep257==0.7.0
pep8==1.7.0
mccabe==0.5.0
boto3==1.3.1